Use syrupy instead of snapshottest for snapshot tests (#14804)
Summary:
This other snapshot library has the same API but doesn't delete all the
other snapshots when you run a test suite with -k, and raises an
exception if the snapshots are missing. Plus, it allows you to customize
the name of the snapshot so we don't need to generate a unique one for
each and every graphql test suite.