Change 'requested' to 'launched' in evaluations ui (#15218)
"Requested" is a confusing word in the evaluations UI. When we display
`1 requested / 1 skipped / 0 discarded` it seems like a partition was
requested, but then skipped. It makes sense to think that way because
the materialize conditions can be overridden by skips- e.g. I could say
that in an evaluation where a materialization condition was met, we
"requested" a run but then had that run discarded or skipped.

The two alternatives we've discussed are "materialized" and "launched".
The problem with "materialized" is that while we kicked off a
materialization, it may fail. The issue with launched is we don't
usually use it with assets- on the other hand we still display `Launched
run xyz` once you click materialize. I think it's an intuitive enough
term and clearer than requested.

<img width="1552" alt="Screenshot 2023-07-11 at 8 32 29 AM"
src="https://github.com/dagster-io/dagster/assets/22018973/fc85720c-1ce1-4e8e-b5c8-39e79aa59ab1">
<img width="1552" alt="Screenshot 2023-07-11 at 8 32 06 AM"
src="https://github.com/dagster-io/dagster/assets/22018973/d9d01b5d-0e55-4b86-858c-72c385ff60b3">