feat(toml): allow code location name to be set from pyproject.toml (#15245)
## Summary & Motivation
Allow the code location name to be specified in `pyproject.toml`, rather
than using the module path of the code location.

This is used in `dagster-dbt project scaffold`, so that we can display
the name of the code location as the name of the dbt project, rather
than as `scaffold.definitions`.

## How I Tested These Changes
pytest, local