[ui] Fix WebWorkers when using path-prefix (#15299)
## Summary & Motivation

Resolves #14894.

WebWorkers are failing in the `--path-prefix` case again. In this case
it surfaced with global search, but the DAG layout worker has been
broken too. This may have been broken for a while.

The magic `__webpack_public_path__` var apparently cannot be prefixed
with a namespace. When the `(self as any)` is removed, Webpack behaves
as expected and the path prefix is correctly applied to the subsequent
`import`.

I added `@ts-expect-error` annotations to this to suppress the resulting
error.

## How I Tested These Changes

With `--path-prefix=/foo`, run the app and perform a search. Verify that
search behaves correctly, with no WebWorker-related errors.

Repeat without path prefix, verify same.