[ui] Hide __ASSET_JOB references on the sensor pages (#15263)
## Summary & Motivation

This is a quick fix for the issue raised in
https://github.com/dagster-io/dagster/issues/14844. We don't have the
information about the exact assets in your selection, but we can see
that the targets are asset jobs and show "A selection of assets". (Fixes
https://github.com/dagster-io/dagster/issues/14844)

I think there are a few other things we should consider changing:

- Backfills have "targets" and sensors have "Asset / Job". I'm tempted
to change the labeling in the table + details page to say "Target" to
match the backfill terminology.

- We are currently showing `sensor.metadata.assetKeys` in the Asset /
Job column. Confusingly these are not the targeted asset keys, they're
the monitored asset keys (if the sensor is an @asset_sensor). To make
this more clear in the code I wrapped the rendering into a component
called `<SensorMonitoredAssets />`, but the placement of them in the
table (beneath the "A selection of assets" or job target info) might be
confusing to users. I don't have a better placement idea - perhaps we
just remove them from the list view.

<img width="1400" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/4c5f2409-62e0-42ab-9ed2-4b1a79f746a0">

Loom of this:
https://www.loom.com/share/056baeb022fe4c95874656b2a576267b

## How I Tested These Changes

I tested these with a few asset_sensor, op job and asset job sensors

---------

Co-authored-by: bengotow <bgotow@elementl.com>