[auto-materialize] Fix mapping from unpartitioned to time partitioned assets (#15368)
## Summary & Motivation

This is a temporary fix for a perf regression caused by removing the
time window partition scope minutes. In the specific case where an
unpartitioned asset is upstream of a time partitioned assets, the
general interpretation that we take is that this update is only relevant
to the latest partition of the time partitioned asset. However, we don't
currently represent this in the

The net result is that the update to the unpartitioned upstream would
result in us evaluating every single downstream, then creating a bunch
of discard reasons for all of the partitions that we skipped (due to the
rate limit). Beyond being slow, this would be confusing to a user who
would not understand why those partitions would need to be discarded in
the first place.

## How I Tested These Changes