Coerce asset_partitions to a set
This comparison hasn't been as quick as the comment indicates. Profiling
suggests that a lot of time is being spent in this particular set
comparsion.

It looks like `asset_partitions` is being provided as an
`Optional[Sequence[AssetPartitionKey]]` but being treated as if it's a
set. This naively coerces the sequence to a set for our comparison but I
think a better implementation might involve updating our call sites to
make sure we're passing a set in the first place.