Filter out default executor config when splatting defaults (#15377)
## Summary

Special case the default multi-process executor config when splatting
out run config yaml, since it can be confusing for users:

```yaml
execution:
  config:
    multiprocess:
      max_concurrent: 0
```

Here, `max_concurrent: 0` is a sentinel value that adjusts concurrency
to match the number of cores on the executing machine. The value of
including this entry in the default config is lower than ops/resources
since most users won't need to touch it - it's more liable to cause
confusion.

## Test Plan

Update unit tests.