dagster-helm rename dagit -> dagster-webserver (#15356)
## Summary & Motivation

Update the Dagster Helm chart with the new "webserver" terminology
instead of "dagit."

- Rename all fields using "dagit" in `values.yaml`:
    - `.dagit` -> `.dagsterWebserver`
    - `.ingress.dagit` -> `.ingress.dagsterWebserver`
    - `.ingress.readOnlyDagit` -> `.ingress.readOnlyDagsterWebserver`
- Default pod name now starts with"dagster-webserver" instead of
"dagster-dagit"
- To manage backcompatibility in templates, a new named template
"dagster.backcompat" was added. This defines a dictionary that mounts
merges of the old and new names at all renamed paths. It is intended to
be "imported" into the variable `$_` and used like so:

```
{{ $_ := include "dagster.backcompat" . | mustFromJson }}
dagster-webserver -h 0.0.0.0 -p {{ $_.Values.dagsterWebserver.service.port }}
```

- In the schema, remove the `dagit` fields entirely. This does not
prevent any values passed in under the old names from being accessed
during template rendering. I am not sure if this is the best approach,
but it keeps our schema simple. The alternative would be to clone the
subschema used for `dagsterWebserver` but mark all the fields optional--
this would be necessary because there are no longer defaults defined for
the old fields in `values.yaml`, so validation fails if partial values
are provided under the old names (which in practice does not matter
because they are merged into the values under the new names).
- Update relevant docs with the new terminology.
- Update integration test suites with the new terminology.

## How I Tested These Changes

I modified some of the schema tests we have to parametrize on old vs new
field names. This verifies that using either name works. I confirmed
that changing the parametrization to use a name that is neither new name
or old name fails.