Update multiprocessing executor max_concurrent sentinel to None (#15389)
## Summary

Based on
https://github.com/dagster-io/dagster/pull/15377#discussion_r1268400399

Currently, the sentinel value for the `max_concurrent` config field in
the default multiprocessing executor is `0`, which indicates that the
value should be dynamically resolved to `multiprocessing.cpu_count()`.
This appears quite confusing in the config editor, since `0` is not
clearly a sentinel without the docstring context.

This PR changes the default value to be `None`, which is now acting as
the sentinel. `0` continues to also operate as the sentinel value.

Docstring + apidocs updated accordingly.

## Test Plan

New unit tests.