[ui] Use a regular RR link for historical alert on AssetView (#15440)
## Summary & Motivation

Resolves #15410

The link in the `HistoricalViewAlert` on `AssetView` is really for
navigation, not for in-place query param updates. Instead of using query
persisted state, just build a link to the appropriate `time` value.

## How I Tested These Changes

Navigate from a run to a historical view of an asset. Verify that when I
click the link in the alert, I navigate via normal link behavior, and
that I can navigate back with the back button, as expected.