feat(dbt): check the type of `dagster_dbt_translator` (#15477)
## Summary & Motivation
`dagster_dbt_translator` should be an instantiated instance of a class.
Validate the argument, and add an example.

We don't attempt to instantiate the argument on behalf of the user if
the passed argument is `Type[T]` where `T = TypeVar("T",
bound=DagsterDbtTranslator)` since the subclassed type might need its
own arguments to instantiate itself.

## How I Tested These Changes
N/A