[auto-materialize][refactor] Remove `bfs_filter_asset_partitions` from auto-materialize logic (#15498)
## Summary & Motivation

This replaces the auto-materialize logic with a single toposorted loop
through the graph. Each asset is processed one by a single call to
`get_auto_materialize_conditions_for_asset`, where we evaluate each
condition in order (first materialize conditions, then skip conditions,
then discard conditions), building up a set of conditions that apply to
each asset, as well as a set of asset partitions we'd like to
materialize for each asset.

One nice benefit here is that we don't end up needing to explicitly
visit the children of assets that we plan to materialize (which can
incur a fair amount of expense in the partition mapping layer). Instead,
the children inspect each of their parents to see if it would even be
possible to materialize at the same time as them, and if not, we
completely ignore that asset.

Beyond that, this makes it significantly simpler to add new conditions,
as there is a single place to place this information.

Finally, we (at long last!!!) accomplish the dream of unifying
freshness-based logic and non-freshness based logic. The freshness
calculation gets evaluated as a simple additional condition that happens
in the same toposorted loop as everything else.

Right now, there's a fair bit of weirdness in the shape of the
AutoMaterializeCondition <> AssetKeyPartitionKey mapping that we pass
around. The shape we really want is AssetKey ->
[AutoMaterializeCondition -> AssetKeyPartitionKey], as this is the end
shape that we end up using inside the AutoMaterializeAssetEvaluation
object, so that's what I have going on in the new functions.
Realistically, get_auto_materialize_conditions_for_asset could also just
return an AutoMaterializeAssetEvaluation directly. A future PR can trim
out a bunch of code that requires us to convert to and from different
formats, just wanted to minimize the surface area of this already-large
PR.

Some tests were changed, I will leave comments on all such changes.

## How I Tested These Changes