[fix] Ensure Airbyte, Fivetran Pythonic resources go through initialization when loading from instance (#15547)
## Summary

Fixes an issue which arose when trying to use `EnvVar`-configured
Pythonic resources in a Definitions-time context (e.g. loading Fivetran
or Airbyte assets from an instance).

This was a regression which occurred as a side-effect of the change in
#14306 which made Pythonic resources expect fully processed config.
#14527 introduced a new `process_config_and_initialize` method which
fully evaluates a Pythonic resource's config, and fixed the DBT
resources, but did not apply the same change to Fivetran and Airbyte.

This PR calls `process_config_and_initialize` as part of the
load-from-instance methods in both integrations, and puts this path
under test. It's unfortunate that this wasn't caught because none of our
existing unit tests hit the overlap of `EnvVar` + Pythonic resource +
Definitions-time load-from-instance call.

## Test Plan

Put this `EnvVar` + Pythonic resource + load-from-instance case under
unit test.