Enforce max_materializations_per_minute as positive (#15555)
Eh. Setting `max_materializations_per_minute=0` or to anything negative
is vague. It currently blocks all materializations, which I guess is
expected, but not useful and potentially confusing. I could imagine
users expecting it to disable the limit (which they can do with
`max_materializations_per_minute=None`, much better).

This is a breaking change, but it's to an unlikely setting on an
experimental api. It'll take a while, but I'm going to check if any
cloud customers have this set. Should lead to a useful script anyway.

**Update**: no cloud orgs have max_materializations < 1