[ui] UnstyledButton (#15415)
## Summary & Motivation

This pattern has come up plenty of times, so I'm creating a component
for it. In cases where we want interactivity (most often a click event)
on an otherwise non-interactive element, we often end up one-offing
unstyled buttons.

This PR introduces an `UnstyledButton` component for these cases. There
are options to increase the click area of the button and to show the
focus outline.

See Storybook preview for examples.

## How I Tested These Changes

Storybook examples.