[ui] Link to auto-materialized assets from Run (#15598)
## Summary & Motivation

Resolves #15225.

For runs that have automaterialize tags, show an "Auto-materialized" tag
in the header. Hovering on the tag shows the assets associated with the
run, and the user can navigate directly to that evaluation ID on the
asset's auto-materialize history page.


https://github.com/dagster-io/dagster/assets/2823852/9ba91cef-d34b-4c5f-8789-d13bb9d75f8c

## How I Tested These Changes

View a run that was launched from an auto-materialize policy. Verify
that the tag appears in the header, and that hovering on the tag shows
the assets assocaited with the run. Verify that clicking to view the
evaluation goes to the asset's auto-materialization history at that
evaluation number.

Verify that long lists of assets are limited by the popover but
scrollable. Verify that asset names are sorted alphabetically.

Adjust the width of the popover to verify that asset names are
middle-truncated if too long for the container.