Partition stale status resolution (#14265)
## Summary & Motivation

Add stale status and cause resolution at partition granularity. The
implementation here is somewhat complicated:

- change the identifier used in the `CachingStaleStatusResolver` to
`AssetKeyPartitionKey` instead of the plain `AssetKey`
- update `DataProvenance` to contain input storage ids (already stored
on `AssetMaterialization` events) in addition to data versions
- given assets A -> B, with A partitioned, the `DataVersion` for A
stored in the provenance tags of a B materialization now uses _the
combined hash of all partitions of A on which B depends_. This is also
used when auto-computing the `DataVersion` of B. This change allows us
to avoid an explosion of tags stored on B materializations in the event
that are many dependency partitions of A.
- Exception to above: If A has >
`SKIP_PARTITION_DATA_VERSION_DEPENDENCY_THRESHOLD` (set to 100 in this
PR) partition dependencies of B, then we revert to old behavior (opt out
of accurate data version computation and stale status reporting for the
A->B edge). This is a temporary limitation that eventually we should
remove.

Given A -> B, with A partitioned, when resolving stale status of B we
use the storage ID of A stored on the last materialization of B as a
cursor:

- check if _any_ new data version record (materialization or
observation) for A has been created after the cursor
- if no, then the dependency hasn't changed, there is no new upstream
data from A, since any dependency partition that was newly materialized
would have to exist post-cursor
- if yes, then for each dependency partition, retrieve two records (1)
the most recent materialization; (2) the most recent materialization
before the cursor. Record (2) corresponds to the materialization last
used as input for B. Compare data versions on these two records-- if
they are the same for at least one dependency partition, then B is stale
(i.e. has new upstream data from A).

Current issues:

- This _does not work_ for observable source assets, since the optimized
method used to fetch the data version tag for a large set of partitions
requires filtering by a single event type (so we can't look for both
`AssetObservation` and `AssetMaterialization` at once).
- The `CachingStaleStatusResolver` currently performs an optimized batch
fetch of the most recent materialization/observation for regular asset
keys, but it implements no such optimization for `AssetKeyPartitionKey`
- We should probably change the provenance tag used for the data version
of a partitioned dependency-- currently the `DATA_VERSION_TAG` is used,
but this is misleading since we are now storing a hash of the data
versions of all dependency assets rather than an actual data version.
- There is no record on the materialization of exactly which upstream
partitions are used to calculate the "combined data version"-- this is
determined at runtime by applying the partition mapping. We might want
to store this information.

## How I Tested These Changes

- New unit tests
- Profiling done with this script:
https://github.com/dagster-io/dagster/pull/15610

Profiling gives a rough estimate of the effect of the dependency limit.
In the benchmark script, resolving stale status for ~2000 partitions
took approximately 8 seconds with no limit and 4 seconds with a limit
(see full results below). This benchmark tests a dead simple scenario
for resolving partition status, and more varied scenarios need to be
explored to better understand performance characteristics here.

```
Experiment settings
┏━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━┓
┃                      Key ┃ Value ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━┩
│           num_partitions │  2000 │
│ override_partition_limit │  True │
└──────────────────────────┴───────┘

Execution times
┏━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━┓
┃ Index ┃                                      Step ┃    Time ┃
┡━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━┩
│     0 │         Resolve StaleStatus of all assets │  0.0443 │
│     1 │ Materialize all 2000 partitions of `root` │ 75.1230 │
│     2 │                 Materialize `downstream1` │  0.1839 │
│     3 │         Resolve StaleStatus of all assets │  7.7065 │
│     4 │                 Materialize `downstream2` │  0.1750 │
│     5 │         Resolve StaleStatus of all assets │  7.5177 │
│     6 │    Materialize single partition of `root` │  0.1642 │
│     7 │         Resolve StaleStatus of all assets │  7.5286 │
└───────┴───────────────────────────────────────────┴─────────┘

Experiment settings
┏━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━┓
┃                      Key ┃ Value ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━┩
│           num_partitions │  2000 │
│ override_partition_limit │ False │
└──────────────────────────┴───────┘

Execution times
┏━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━┓
┃ Index ┃                                      Step ┃    Time ┃
┡━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━┩
│     0 │         Resolve StaleStatus of all assets │  0.0440 │
│     1 │ Materialize all 2000 partitions of `root` │ 74.4319 │
│     2 │                 Materialize `downstream1` │  0.2226 │
│     3 │         Resolve StaleStatus of all assets │  4.0985 │
│     4 │                 Materialize `downstream2` │  0.1963 │
│     5 │         Resolve StaleStatus of all assets │  4.1553 │
│     6 │    Materialize single partition of `root` │  0.1618 │
│     7 │         Resolve StaleStatus of all assets │  4.2651 │
└───────┴───────────────────────────────────────────┴─────────┘
```