GQL stale status partition resolvers (#15657)
## Summary & Motivation

- Rename `AssetNode.currentDataVersion` -> `dataVersion`. (This field is
not currently being called from Dagster UI so its a safe rename).
- Make `CachingStaleStatusResolver` return an `StaleStatus.FRESH`/`[]`
when requesting the status/causes for a partitioned asset without
specifying a partition key (status quo throws an error)
- Add partition-scoped resolution for GQL `AssetNode` fields:
    - `staleStatus`
    - `staleStatusCauses`
    - `dataVersion`

Two changes for each of the above fields:

- The singular resolver (e.g. `resolve_staleStatus`) now takes an
optional `partition` argument specifying a single partition key. An
error is thrown if this is specified for a non-partitioned asset.
- A new `XByPartition` field(e.g. `staleStatusByPartition`) field is
added that takes on optional `partitions` argument, which is a list of
partition keys. It returns the requested field in a list corresponding
to the order of the passed partition keys. If no list of partition keys
is specified, the result for all partition keys is returned (note that
this form requires you to know in advance the order of all partition
keys to be interpretable).

## How I Tested These Changes

New unit test