[ui] Allow filtering the global graph by asset group (#15672)
## Summary & Motivation

Fixes: https://github.com/dagster-io/dagster/issues/15355

This PR enables filtering the global asset graph by asset group, and
brings asset group multi-selection to the global asset catalog.

<img width="657" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/37f90294-35c2-4ab0-b598-f43f4f596f97">

I considered porting the fancy RunsFilter UI over to the asset catalog,
but at the moment group is the only available filter, and the RunsFilter
UI exposes a "Filter" button and a separate list of "filter tags", which
we don't have a great place to show in the asset graph. I think that
having a single component with typeahead filtering and multi-selection
may be best until we redesign this UI a bit more significantly.

## How I Tested These Changes

I tested this in both the global asset graph and the asset catalog in a
repo that contains many groups, some of which have overlapping names. I
also verified that clicking around the global graph does not clear your
filter, and that reloading the pages preserves / rehydrates the filter
state correctly.

---------

Co-authored-by: bengotow <bgotow@elementl.com>