[ui] Leave existing LS keys in place (#15700)
## Summary & Motivation

Don't delete the localStorage keys we're migrating away from, just to be
safe. OSS users may need to revert to an older version, or we may have
to roll back a Cloud deploy for some reason. In either case, users who
load the app will have lost their old state, which would be an
unfortunate experience.

We can always clean these up later if we want, but in the meantime
they'll just be orphaned keys on the user's localStorage.

## How I Tested These Changes

Buildkite