[ui] Show freshness policy info in Fresh state, use "minutes old" not "minutes overdue" (#15481)
## Summary & Motivation

This PR addresses a few pieces of feedback from @sryza:

- Hovering over the "Fresh" icon opens the panel showing the calculation
of the data age - it's no longer restricted to assets that are overdue.

- The panel (and all popover panels) have a slightly stronger shadow.
(Discussed this with Josh a bit and we don't want to go too dark, but
this does look better)

This PR also fixes a bug I hadn't noticed earlier. The final text
essentially says "This asset's data is X minutes old, and the policy
allows it to be Y minutes old", but it was using the
"currentMinutesLate" which is not "minutes old" but "minutes overdue". I
exposed the "minutes old" value in GraphQL and updated the copy.

- Todo: I think it'd be nice to rename `currentMinutesLate` to
`currentOverdueMinutes` for consistency with `currentLagMinutes` and
`maximumLagMinutes`, but that would make this a bit messy.

## How I Tested These Changes

Updated python tests + JS storybooks

---------

Co-authored-by: bengotow <bgotow@elementl.com>