[ui] Display stale status / causes on individual asset partitions (#15688)
## Summary & Motivation

This PR stacks on https://github.com/dagster-io/dagster/pull/15657 and
adds Stale tags to the Asset > Partition details view. This allows you
to see if a specific partition needs to be re-materialized after the
upstream partitions it incorporates were changed.

Right now, this value is fetched for the specific partition shown and is
shown next to the green "Materialized" tag. We want to keep "Green" in
this UI for now, since we can't fetch this info in bulk fast enough to
put filterable stale / yellow dots in the left sidebar or asset health
bar.

## How I Tested These Changes

I updated the storybooks for this to show a stale cause.

I also took a partition mapped asset, materialized a partition,
re-materialized one of it's upstreams, and observed that the partition
appeared "Stale" while others did not.

I also fixed a small bug I discovered where the UI would stick in a
loading state if the only attempted run for a partition failed to
materialize and there is no `latest` event.



![image](https://github.com/dagster-io/dagster/assets/1037212/e9bb2d14-617f-4315-a3ed-796360e18d35)

---------

Co-authored-by: bengotow <bgotow@elementl.com>