Skip integration tests in unit test builds
Over time, our test suite has evolved to have a mix of both unit and
integration tests. This represents a blocker to running unit tests on
contributor PRs because we don't want to expose the ability to interact
with authenticated 3rd party systems.

This build shows what fails if we don't provide any secrets or IAM
credentials to buildkite:

https://buildkite.com/dagster/isolated-dagster/builds/18

A lot of these source back to this `get_aws_creds()` function. For now,
I'm going to annotate with a pytest skip. Eventually, we might want to
explicitly pull these out into their own integration build or rewrite
them to not integrate with real systems.