Add `config` argument to graph_asset (#15384)
## Summary & Motivation

Resolves #14944.
Resolves #12647.

- Add `config` arg to `@graph_asset`
- Add `config` docstring to `@graph`
- Light refactor of `@graph_asset` higher-order decorator style to
eliminate duplication of long argument list.

`@graph_asset` is currently missing `config`. This has been reported by
multiple users. IMO it should clearly be present since it's part of the
standard functionality of `@graph`.

@sryza in #12647 you wrote:

>@asset doesn't currently have a config attribute, but we've been
discussing adding one. Arguably, @graph_asset and @asset should be
consistent on this.

`@graph_asset` combines arguments of both `@graph` and `@asset`, so I
think it is important to add this regardless of whether it is added to
`@asset` as well, as we need to be consistent with `@graph`.

## How I Tested These Changes

New unit test.