[auto-materialize][bug] Fix unpartitioned to partitioned rematerialization logic (#15725)
## Summary & Motivation

Previously, rematerializing an unpartitioned asset upstream of a time
(or dynamic) partitioned asset would not result in the latest partition
of that asset being rematerialized, as we would never register that
unpartitioned as "updated" with respect to its parents. Due to this, we
would only ever materialize the new partition of the upstream asset when
it was missing.

## How I Tested These Changes

Added a new test, and updated two old ones to reflect this corrected
behavior.