Handle partitions natively in W&B IO Manager (#15170)
## Summary & Motivation

This PR aims to enrich the W&B integration by enhancing the
functionality and user experience of the IO manager. We've focused on
native handling of (multi-)partitioned assets, which streamlines the
process and ensures our implementation is more robust. We also addressed
and resolved several bugs, to provide a more seamless experience to
users.

In our commitment to clarity and ease of use, we've added comprehensive
examples showcasing the use of partitions and multi-partitions, as well
as how to specify a version or alias in the configuration metadata.

We’ve also given much-needed attention to user-facing errors, rewording
them for better understanding and visibility. Moreover, we introduced
additional errors to alert users when there might be issues with their
provided configuration, improving their ability to self-diagnose and
rectify issues.

## How I Tested These Changes

The examples added were run in multiple scenarios to ensure that they
cover common use cases and successfully demonstrate the intended
functionality.

We also replicated common bugs that were reported, and confirmed that
the fixes eliminated these issues.

Lastly, we ensured backward compatibility and non-disruptive deployment
with integration tests.

We're excited to share these updates and we believe these improvements
will significantly enhance the user's experience. Looking forward to
your feedback, Dagster team!

---------

Co-authored-by: yuhan <yuhan@elementl.com>