[dagster-fivetran] Add additional options (e.g. poll interval, timeout) to load_from_instance method (#15625)
## Summary

It is often useful to configure a maximum sync duration for a Fivetran
sync, especially in the case it is rescheduled (at which point the sync
is resumed a day later once API limits are restored).

https://fivetran.com/docs/getting-started/syncoverview

Allows users to pass these options, previously only found on
`build_fivetran_assets`, to `load_assets_from_fivetran_instance`.

Unfortunately, I'm not sure we have a good way of handling the
rescheduled case outside of this - there is no API endpoint to cancel a
sync (in case we wanted to avoid rescheduled state entirely). Our only
options seem like "poll until reschedule starts" (current behavior) or
"mark failed on reschedule" (in which case Dagster misses the
rescheduled run).

## Test Plan

Pass along args to unit test.