move "dbt models and Dagster software-defined assets" from tutorial t… (#15683)
…o reference

## Summary & Motivation

The beginning of the dbt tutorial currently includes a section titled
"dbt models and Dagster software-defined assets", which explains the
parallels between dbt models and SDAs. This is great content, but I
think it's a lot to digest at once for someone who's opted for the
tutorial because they want a gentle walkthrough. It exposes them to
these in a short span: the Python API for defining an asset, the term
"Software-defined asset", the term "operation", and the term "asset
dependency".

This PR proposes moving this content to the reference section of the dbt
docs.

## How I Tested These Changes