[ui] LS key migration fixes (#15813)
## Summary & Motivation

Try to fix a couple of possible issues with localStorage key migration.

- Since we're not deleting old keys (because the user might revert to an
old version, or if we were to have to revert on Cloud), we're
effectively doubling the storage used by the affected key-value pairs.
This means users might hit quota errors on write. Wrap the `setItem` in
a try-catch so that these are less noisy.
- If the new key already exists, we're already done -- don't waste time
writing it again. (I was thinking that this could be an issue in the
revert case, since the revert state could end up overwriting the new-key
state. But I think that's not actually a problem, since anything being
done in the revert state is probably intended and *should* overwrite any
new keys once the user is able to move forward with the new version.)
- Add a `deleteExisting` flag to the migration function. This gives me a
switch to flip once we're ready to clean up old keys.

## How I Tested These Changes

Jest