Brownfield essentials dagster-external protoype (#15760)
## Summary & Motivation

Internal companion PR: https://github.com/dagster-io/internal/pull/6440

Prototype implementation of `dagster-external` library used for
separation of orchestration and transform layers.

New resource `ExternalExecutionResource` works similarly to
`DbtCliResource`. Create an asset with this resource and you can call
`run` on the resource to execute an arbitrary program (via `Popen`). The
resource pattern allows for the user to pull arbitrary
(JSON-serializable) data off the `OpExecutionContext` and
`DagsterInstance`, which is packaged together with other standard info
(e.g. `run_id`) and sent over the wire to be reconstructed as an
`ExternalExecutionContext` on the other side.

The external process should import `dagster_external` and call
`init_dagster_external` (as of this PR, only Python is supported but use
of `dagster-external` as a CLI is coming soon). This will read env vars
injected by the spawning orchestration process, which are used to
establish communications with the orchestration process. As of this PR
stdio, a user-provided FIFO, or an automatically created temporary FIFO
are IPC options (they can be varied independently for input and output
to the external process). Socket support is planned.

`init_dagster_external` returns a singleton instance of
`ExternalExecutionContext`-- after initialization, the singleton can
also be retrieved by calling `ExternalExecutionContext.get()`.

The `ExternalExecutionContext` has a variety of accessors for the
serialized context information sent from the orchestration process, as
well as methods to stream data back to the orchestration process. This
PR supports only a small number of such "reporter" methods and the names
are not set in stone, this will be iterated on in the near future.

## How I Tested These Changes

New unit tests.